requirements.txt
# ─── J.A.R.V.I.S. Dependencies ─────────────────────────────
# Core AI
openai>=1.30.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
websockets>=12.0

# Voice
edge-tts>=6.1.0
SpeechRecognition>=3.10.0
pygame>=2.5.0
PyAudio>=0.2.14

# System Control
psutil>=5.9.0
pyautogui>=0.9.54
pyperclip>=1.8.0
pycaw>=20240210
comtypes>=1.3.0
pywin32>=306  # Windows only — for window automation

# WhatsApp
pywhatkit>=5.4

# ESP32 / IoT
pyserial>=3.5
paho-mqtt>=2.0.0

# Image & Vision
aiohttp>=3.9.0
Pillow>=10.0.0

# PDF & Documents
PyPDF2>=3.0.0
pytesseract>=0.3.10

# QR Codes
qrcode[pil]>=7.4

# Browser Automation (optional)
# selenium>=4.15.0

# Environment
python-dotenv>=1.0.0

# Utilities
python-multipart>=0.0.9

